+++ R/geom-transformers.R
@@ -1123,12 +1123,13 @@ st_line_sample = function(x, n, density, type = "regular", sample = NULL) {
 } #nocov end
 
 #' @param dist numeric, vector with distance value(s)
-#' @name st_project_point
+#' @name st_line_project_point
+#' @returns `st_line_interpolate` returns the point(s) at dist(s), when measured along (interpolated on) the line(s)
 #' @export
 #' @examples 
-#' st_interpolate_line(st_as_sfc("LINESTRING (0 0, 1 1)"), 1)
-#' st_interpolate_line(st_as_sfc("LINESTRING (0 0, 1 1)"), 1, TRUE)
-st_interpolate_line = function(line, dist, normalized = FALSE) {
+#' st_line_interpolate(st_as_sfc("LINESTRING (0 0, 1 1)"), 1)
+#' st_line_interpolate(st_as_sfc("LINESTRING (0 0, 1 1)"), 1, TRUE)
+st_line_interpolate = function(line, dist, normalized = FALSE) {
 	stopifnot(inherits(line, "sfc"), all(st_dimension(line) == 1), 
 		is.logical(normalized), length(normalized) == 1,
 		is.numeric(dist))
@@ -1136,6 +1137,6 @@ st_interpolate_line = function(line, dist, normalized = FALSE) {
 		message_longlat("st_project_point")
 	line = st_cast(line, "LINESTRING")
 	recycled = recycle_common(list(line, dist))
-	st_sfc(CPL_interpolate_line(recycled[[1]], recycled[[2]], normalized), 
+	st_sfc(CPL_line_interpolate(recycled[[1]], recycled[[2]], normalized), 
 		   crs = st_crs(line))
 }